Move similar chunks into the LineEditor trait
I am using the trait here as an interface that is so popular in a lot of
OO langs. I am not super keen on using this I would prefer if there was
a way to get this trait changed into a struct and modify the current
ViLineEditor and EmacsLineEditor to adhere a common interface that we
will plug in that LineEditor struct. (And obviously we need to call it
Reedline)