Revamp Prompt states
The idea here is that what we want to display on the screen which can
impact the prompt should be available as part of the prompt's interface.

So here are the things that can impact what we show with our prompt
- The line-editor can either be in Emacs mode or VI mode (for now)
  [These are exclusive to one another]
- The line-editor can be either in editing a command or searching
  history mode [These two states are also exclusive]
- Lastly the command can be either single-line or multi-line.

Now all these variables can happen togeter (i.e. they for a cross
prodcut)
- example 1: You can be in vi-normal mode and be editing history search
  term and that particular term earlier resulted in a multi-line
  command. (So, we need to showcase all these three variations together)

one way to do this is using a struct. I found that it was making the
more frequent case (of being in edit mode) overly complex so I chose to
expose the functionality as separate methonds.