Remove bool from remember_undo_state
Having a function do two different things based on a boolean that is
passed usually results int unreadable and unmaintable code and this was
no different in my opinion.

Now the users of this interface don't have to worry about what that
boolean flag is.
Given that Rust does not have kwargs the readability of passing booleans
into a function results in having to read the interface and
documentation (or implementation) of the particular function. Which we
can also avoid if we don't do it.
Lastly removes possibility of creating a bug because a wrong boolean
value was selected.