Attempt fix of `ClearScrollback`
Clearing the scrollback did not work on most terminal emulators.
By reordering two ANSI control sequences, this seems to work on Linux
with most terminal emulators.
Due to a missing implementation of the exact details this will probably
not clear the scrollback under the old Windows API.

Related to the discussion in #120 and nushell/nushell#5089