Fix vi-mode word motions (#425)
* Fix vi-mode word motions

The 'w' motion goes to the beginning of the next word to the right, not
the end. The 'e' motion, which is introduced here, goes to the end, but
lands on the last character, not after it.

* Add test case for `move_word_right_end`

Ensure Unicode safety as well as that the cursor will move to the next
word end when already at a word end

Co-authored-by: sholderbach <sholderbach@users.noreply.github.com>