Support vi mode ; and , motions
The f, F, t, and T left-right motions can be repeated with the ; motion,
or reversed with the , motion:

> ;    Repeat latest f, t, F or T [count] times. See cpo-;
>
> ,    Repeat latest f, t, F or T in opposite direction
>      [count] times. See also cpo-;

I named these "to-till" motions as the first word of the description for
the f and T motions are "To" and "Till" respectively.

Add storage for the last To or Till motion on the `Vi` struct, and pass
it through to `parse_command()` so `;` or `,` can re-use it.

When a to-till motion is fully parsed the new to-till motion is stored
in the `Vi` struct.