Correctly implement ,
The last to or till motion should be recorded from f, F, t, or T.
Using , to reverse the to/till motion must not change the motion that
will be replayed.

In other words, if you f to move to the right then pressing , will only
move you left.  Pressing , will not swap between the two matching
characters your cursor is on or between.