make control-w behave like bash
'man bash' documents this as:

       unix-word-rubout (C-w)
              Kill the word behind point, using white space as a word boundary.
              The killed text is saved on the kill-ring.

zsh does not have a concept of small/big words, but its 'word' is slightly
bigger than nushell's small word, because it includes -./_ and a few other
punctuation marks.

fish's control-w is slightly larger than nushell's small word in that it
seems to delete up to the next / (and doesn't need an extra control-w to
delete the '/' 'thing/').