Simplify to just `with_buffer_editor`
The old `with_buffer_editor` would have not set the file and
`with_buffer_editor_command` had unclear semantics if you should provide
the file to the command or not.

Instead of going the route of deprecation let's just make the breaking
change and update `with_buffer_editor` to a more practical signature.
Expect a `Command` and a `PathBuf` for the file.
If the file is not among the args push it.