Remove questionable `unsafe impl Send`
`HistoryCompleter` is not `Send` due to it containing a reference to a
`&dyn History` which was only `trait History: Send`.
Here the property of a type T needing to implement `Sync` for its `&T`
to be `Send` comes into play. Thus this `unsafe impl Send` is unsound.

The change of making `trait History: Send + Sync` makes
`HistoryCompleter: Send` automatically and should be compatible for all
sound implementations.