feat(vi-mode): add `gg` and `G` motion
There is a bit of divergence in behaviour of `gg` and `G` motion of
reedline and vim. In vim, these motion maintain the column position of
the cursor, but reedline doesn't. Instead `gg` takes the cursor to the
very first character and `G` takes to the very last character.

Also when using these motion with a command, e.g. `c`, `d` and `y`,
there are some inconsistency in handling the new line character due
reedline not supporting line wise copy unlike vim.

Added test case for `dgg`, `dG`, `cgg` and `cG` command.