feat(vi-mode): add `gg` and `G` motion
There is a bit of divergence in behaviour of `G` motion of reedline and
vim. In vim, `G` takes the cursor to the first character of the last
line. In reedline, it tales to the very last character of the buffer
instead.

Also when using these motion with a command, e.g. `c`, `d` and `y`,
there are some inconsistency in handling the new line character.

Added test case for `dgg`, `dG`, `cgg` and `cG` command.