fix(vi-mode): fix `^` motion
Previously the `^` motion was behaving like the `0` motion which is to
move the cursor to the start of the line. But in vim, the `^` motion is
different from `0` motion. In vim `^` moved the cursor to the first
non-blank character of the line. (See `:h ^`)

EDGE CASE: When the cursor is on the first line with no indentation, the
`^` motion takes to the start of second word.
The workaround is to use `gg` motion in this particular context.

Disabled test case for `d^` and `c^` command since it doesn't have its
old behaviour and the new correct behaviour doesn't seem plausible to
achieve trivially.